Update put_assoc cheatsheet to use changesets
The existing docs use update_in on a struct and pass that to put_assoc
which doesn't actually perform any updates because:

"structs are not change tracked in any fashion.
In other words, if you change a comment struct and give it to put_assoc/4,
the updates in the struct won't be persisted. You must use changesets instead."