erase from persistent_term without blocking registry
This was an interesting problem. It started with me updating the
dependencies in phoenix_live_dashboard and noticing test failures related
to our test of the Ecto Stats page. The stats page calls `Ecto.Repo.all_running`
to get a list of repos. The problem was that the test case still got
an old repo as running, even though the previous test already stopped
it (using start_supervised). It turns out that the :persistent_term.erase
was slowing the registry GenServer down too much, therefore it wasn't
processing the DOWN messages fast enough and not deleting the old repo
entries from ets before the new test was already running.